Test code via GitHub actions
This change adds an initial configuration for GitHub actions so that
it's possible to check `Test::CheckManifest` on multiple Perls as well
as multiple platforms.  This config will check any pushes to any branch
as well as all pull requests.

It turns out that to test multiple Perls, it's easier to do this on
Linux and this has a separate configuration to Windows and MacOS because
in the Linux case one can use Docker images which means that the
instances are set up and running very quickly.

In the case of Windows and MacOS it's necessary to install `Dist::Zilla`
which takes some time (even with tests turned off), hence it doesn't
make any sense to check multiple Perls on these systems because it'd
take too long (and I'm not sure if so many Perls are available on those
systems anyway).

Even though the syntax of this module is compatible with Perl 5.6 and
upwards, it's only possible to test 5.20 and upwards as these are the
only docker images available which have `Dist::Zilla` pre-installed.